--- src/server/server.rs
@@ -1,2 +1,0 @@
-mod process;
-pub mod supervisor;
